backend/requirements.txt
flask==2.3.3
flask-socketio==5.3.6
flask-cors==4.0.0
flask-jwt-extended==4.5.3
pymongo==4.6.1
python-dotenv==1.0.0
openai==1.3.7
python-multipart==0.0.6
bcrypt==4.0.1
gunicorn==21.2.0
eventlet==0.33.3 